requirements.txt
django-admin-interface
pillow
django-rest-framework